Move pop-out.component into platform ownership
