+++ apps/browser/src/platform/popup/components/pop-out.component.html
